Move the main schemars_0_8 test file to a subdirectory
Move the utils code as well.
